+++ R/summarize_test.R
@@ -39,10 +39,28 @@
 #'
 #' @examples
 #'
-#' # TO PREPARE
-#' x <- 1
-#' \donttest{
-#' }
+#' mod <-
+#' "
+#' m ~ a*x
+#' y ~ b*m + x
+#' ab := a * b
+#' "
+#'
+#' mod_es <- c("y ~ m" = "l",
+#'             "m ~ x" = "m",
+#'             "y ~ x" = "n")
+#'
+#' sim_only <- power4test(nrep = 2,
+#'                        model = mod,
+#'                        pop_es = mod_es,
+#'                        n = 100,
+#'                        do_the_test = FALSE,
+#'                        iseed = 1234)
+#'
+#' test_out <- power4test(object = sim_only,
+#'                        test_fun = test_parameters)
+#'
+#' test_out
 #'
 #' @export
 summarize_tests <- function(object) {
@@ -62,6 +80,10 @@ summarize_tests <- function(object) {
 #' object or a `test_summary` object,
 #' depending on the print method used.
 #'
+#' @param digits The numbers of digits
+#' after the decimal when print
+#' numeric results.
+#'
 #' @param ... Optional arguments.
 #' Not used.
 #'
@@ -74,16 +96,103 @@ summarize_tests <- function(object) {
 #'
 #' @export
 print.test_summary_list <- function(x,
+                                    digits = 3,
                                     ...) {
-  print("WIP. Not ready")
+  for (xx in x) {
+    print(xx,
+          digits = digits,
+          ...)
+  }
+  return(x)
 }
 
 #' @rdname summarize_tests
 #'
 #' @export
 print.test_summary <- function(x,
+                               digits = 2,
                                ...) {
-  print("WIP. Not ready")
+  x0 <- unclass(x)
+  nvalid <- x0$nvalid
+  test_attr <- x0$test_attributes
+  test_name <- test_attr$test_name
+  cat(header_str(paste0("<", test_name, ">"),
+                 hw = .7,
+                 prefix = "\n",
+                 suffix = "\n\n"))
+  if (is.vector(x0$mean)) {
+    results_type <- "vector"
+  } else {
+    results_type <- "data.frame"
+  }
+  if (results_type == "vector") {
+    nvalid_sig <- unname(nvalid["sig"])
+    propvalid_sig <- unname(nvalid_sig / x0$nrep)
+    cat("Mean(s) across replication:\n")
+    out0 <- formatC(x0$mean,
+                    digits = digits,
+                    format = "f")
+    print(as.data.frame(rbind(out0)),
+          row.names = FALSE)
+    cat("\n")
+    cat("- The value 'sig' is the rejection rate.\n")
+    cat("- If the null hypothesis is false, this is the power.\n")
+    cat("- Number of valid replications for rejection rate:",
+        nvalid_sig, "\n")
+    cat("- Proportion of valid replications for rejection rate:",
+        formatC(propvalid_sig, digits = digits, format = "f"),
+        "\n")
+  }
+  if (results_type == "data.frame") {
+    nvalid_sig <- unname(nvalid[, "sig", drop = TRUE])
+    propvalid_sig <- unname(nvalid_sig / x0$nrep)
+    one_nvalid <- all(nvalid_sig == max(nvalid_sig))
+    cat("Mean(s) across replication:\n")
+    df_means <- x0$mean
+    if (!one_nvalid) {
+      df_means$p_sig <- propvalid_sig
+    }
+    df_means <- format_num_cols(df_means,
+                                digits = digits)
+    print(df_means)
+    cat("\n")
+    cat("- The column 'sig' shows the rejection rates.\n")
+    cat("- If the null hypothesis is false, the rate is the power.\n")
+    if (one_nvalid) {
+      cat("- Number of valid replications for rejection rate(s):",
+          max(nvalid_sig), "\n")
+      cat("- Proportion of valid replications for rejection rate(s):",
+          formatC(max(propvalid_sig), digits = digits, format = "f"),
+          "\n")
+    } else {
+      cat(strwrap(paste0("- The column 'p_sig' shows proportions of ",
+                        "valid replications for rejection rates."),
+                  exdent = 2), sep = "\n")
+    }
+  }
+  return(x)
+}
+
+#' @noRd
+format_num_cols <- function(object,
+                            digits = digits) {
+  tmp <- as.data.frame(object)
+  tmp2 <- lapply(tmp,
+                  function(x,
+                            digits) {
+                    if (is.numeric(x)) {
+                      x <- formatC(x,
+                                    digits = digits,
+                                    format = "f")
+                      return(x)
+                    } else {
+                      return(x)
+                    }
+                  },
+                  digits = digits)
+  tmp2 <- as.data.frame(tmp2)
+  class(tmp2) <- class(object)
+  tmp2
 }
 
 #' @noRd